+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### To be Released
 
+* feat(addons): add aliases for database user management (i.e. using format: `database-users-<action>`) ([PR#1031](https://github.com/Scalingo/cli/pull/1031))
+
 ### 1.30.0
 
 * feat(addons): add user management commands ([PR#1019](https://github.com/Scalingo/cli/pull/1019))
@@ -11,7 +13,7 @@
 * feat: add database maintenance listing with the new `database-maintenance-list` command ([PR#982](https://github.com/Scalingo/cli/pull/982))
 * feat(addons): add maintenance windows manipulation with the new `addon-config` command ([PR#955](https://github.com/Scalingo/cli/pull/955))
 * feat(install.sh): verify the archive checksum ([PR#988](https://github.com/Scalingo/cli/pull/988))
-* feat(region): more debug logs  ([PR#1007](https://github.com/Scalingo/cli/pull/1008))
+* feat(region): more debug logs  ([PR#1007](https://github.com/Scalingo/cli/pull/1007))
 * fix(events): `link_scm` data types
 
 ### 1.29.1